requirements.txt
# Core dependencies
requests>=2.28.0
aiofiles>=23.0.0
pyyaml>=6.0.0

# API Server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Authentication
pyjwt>=2.8.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0